Adicionada a funcionalidade de avaliar o livro da estante. E exibir a nota geral e os leitores do livro cadastrado.

+++ imperativo/estante.struct.cpp
@@ -7,6 +7,6 @@ struct Estante {
     int idUsuario;
     struct Usuario usuario;
     struct Livro livro;
-    int nota;
+    float nota;
     int situacao;
 };
